refactor: manager and add lib PhoneNumberField
